+++ bitcoin/chainparams.h
@@ -25,6 +25,7 @@ struct chainparams {
 	const char *cli_args;
 	/* The min numeric version of cli supported */
 	const u64 cli_min_supported_version;
+	const char *rpc_args;
 	const struct amount_sat dust_limit;
 	const struct amount_sat max_funding;
 	const struct amount_msat max_payment;